new player command options
