+++ discord_bot/__init__.py
@@ -1,4 +1,4 @@
-from .discord_bot_util import err_embed, info_embed, get_rank_icon_url, set_true_autocompletion
+from .discord_bot_util import err_embed, info_embed, request_submitted_embed, get_rank_icon_url, set_true_autocompletion
 from .discord_commands.command_drops import days_drops_autocompletion, cb_drops
 from .discord_commands.command_player import opt_player_autocompletion, cb_player
 from .discord_commands.command_spin import spin_cb
